Adding/Removing Entities and Components UI I think this will be my last update on this for a very long time. Now is 2024 Sept. 7th midnight, I just came back from the Unreal Fest 2024 Shanghai.
I felt so much energy in Unreal community China, so little in Ubisoft Shanghai, I must move on now. So I have to ditch this project and the AI courses for some time as I need to work on Unreal Engine in the coming years to ship at least a game all by my one. Or make it playable. This is vital for me to move on in the future. Thanks this project and time spent with Cherno, it has been a fun ride and I am sure I am not a person for Engine or Low-level programming :-)